Address review: extract resolveQuestionId and strengthen tests
