+++ frontend/src/embedding-sdk-bundle/components/private/SdkAdHocQuestion/SdkAdHocQuestion.tsx
@@ -2,11 +2,11 @@ import { useMemo } from "react";
 
 import type { SdkQuestionProps } from "embedding-sdk-bundle/components/public/SdkQuestion/SdkQuestion";
 import { SdkQuestion } from "embedding-sdk-bundle/components/public/SdkQuestion/SdkQuestion";
-import type { SdkQuestionId } from "embedding-sdk-bundle/types/question";
 import { deserializeCard, parseHash } from "metabase/common/utils/card";
-import * as Urls from "metabase/urls";
 
 import type { QuestionMockLocationParameters } from "../SdkQuestion/context";
+
+import { resolveQuestionId } from "./utils";
 
 interface SdkAdHocQuestionProps {
   questionPath: string; // route path to load a question, e.g. /question/140-best-selling-products - for saved, or /question/xxxxxxx for ad-hoc encoded question config
@@ -103,25 +103,6 @@ export const SdkAdHocQuestion = ({
     </SdkQuestion>
   );
 };
-
-/**
- * Derives the questionId from URL slug and deserialized card.
- * Returns "new-native" for hash-encoded native cards (e.g. Metabot SQL editor navigation),
- * a numeric/string ID for saved questions, or null for new notebook questions.
- */
-export function resolveQuestionId(
-  slug: string | undefined,
-  deserializedCard: { dataset_query?: { type?: string } } | undefined,
-): SdkQuestionId | null {
-  const extractedId = Urls.extractEntityId(slug) ?? null;
-  if (extractedId !== null) {
-    return extractedId;
-  }
-  if (deserializedCard?.dataset_query?.type === "native") {
-    return "new-native";
-  }
-  return null;
-}
 
 /**
  * This generates route parameters based on the provided URL path
